[CLEAN] FileManagement.java -> FileManipulation
